requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.8.0
asttokens==3.0.0
attrs==24.3.0
autogen-agentchat==0.2.40
beautifulsoup4==4.12.3
blinker==1.9.0
bs4==0.0.2
cachetools==5.5.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
cryptography==44.0.0
cycler==0.12.1
diskcache==5.6.3
distro==1.9.0
docker==7.1.0
docopt==0.6.2
entrypoints==0.4
exceptiongroup==1.2.2
executing==2.1.0
Faker==33.3.0
favicon==0.7.0
FLAML==2.3.3
fonttools==4.55.3
gitdb==4.0.12
GitPython==3.1.44
greenlet==3.1.1
h11==0.14.0
htbuilder==0.7.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
Jinja2==3.1.5
jiter==0.8.2
Js2Py==0.74
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
lxml==5.3.0
Markdown==3.7
markdown-it-py==3.0.0
markdownlit==0.0.7
MarkupSafe==3.0.2
matplotlib==3.10.0
mdurl==0.1.2
more-itertools==10.5.0
msal==1.31.1
narwhals==1.21.1
numpy==1.26.4
openai==1.59.4
packaging==24.2
pandas==2.2.3
pillow==11.1.0
pipwin==0.5.2
plotly==5.24.1
prometheus_client==0.21.1
protobuf==5.29.2
pure_eval==0.2.3
pyarrow==18.1.0
PyAudio==0.2.14
pycparser==2.22
pydantic==2.10.4
pydantic_core==2.27.2
pydeck==0.9.1
Pygments==2.19.1
pyjsparser==2.7.1
PyJWT==2.10.1
pymdown-extensions==10.14
pyparsing==3.2.1
PyPrind==2.11.3
pySmartDL==1.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pywin32==308
PyYAML==6.0.2
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
safehttpx==0.1.6
semantic-version==2.10.0
semver==3.0.2
sentry-sdk==2.19.2
shellingham==1.5.4
simple-websocket==1.1.0
six==1.17.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
SpeechRecognition==3.13.0
SQLAlchemy==2.0.36
st-annotated-text==4.0.1
st-theme==1.2.3
stack-data==0.6.3
starlette==0.41.3
streamlit==1.41.1
streamlit-camera-input-live==0.2.0
streamlit-card==1.0.2
streamlit-embedcode==0.1.2
streamlit-extras==0.5.0
streamlit-faker==0.0.3
streamlit-image-coordinates==0.1.9
streamlit-keyup==0.2.4
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==2.5.5
tenacity==9.0.0
termcolor==2.5.0
tiktoken==0.8.0
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
urllib3==2.3.0
validators==0.34.0
watchdog==6.0.0
wsproto==1.2.0